[#102] ImagePicker 파일 분리
